Changed GUI slider for video volume to be in range 0.0 - 1.0 (was 0 - 10 before, which made it not useable, because OF expects 0.0 - 1.0)

+++ src/quad.h
@@ -155,7 +155,7 @@ public:
     bool bHighlightCenter;
     bool bHighlightRotation;
 
-    int videoVolume;
+    float videoVolume;
     float thresholdGreenscreen;
     int bgSlideshow;
     unsigned int currentSlide;
